requirements.txt
requests~=2.32.3
requests-toolbelt~=1.0.0
SQLAlchemy~=2.0.40
fastapi~=0.115.12
pillow~=11.2.1
passlib~=1.7.4
pydantic~=2.11.4
aiofiles~=24.1.0
jwt~=1.3.1
PyJWT~=2.10.1